Readme for the w3c/epub repository.
This is to act upon https://github.com/w3c/epub-specs/issues/2619. It is, essentially, giving a logical entry point to the outside world in finding the EPUB related GitHub entries.

In contrast to what the text says, the repository is still open to issues, in order to handle this PR. Once we get to a consensus as for the content, and this PR gets merged, issues will be disallowed.